Wrapper to start app with default db
